requirements.txt
torch
torchvision
matplotlib
numpy
scikit-learn
pillow
tqdm
jupyter
uvicorn